hard coded player, door, and boon spawn; implemented rudimentary paint function to display the current board
